Reimplemented the way depth is calculated to use only unspliced alignments.  Also reworked some of the "samtools depth" code into our DepthParser class to allow for very fast and low memory per base coverage calculation.

+++ include/seq_utils.hpp
@@ -41,7 +41,7 @@ class SeqUtils {
     
 public:
     
-    static uint32_t hammingDistance(const string& s1, const string& s2) {
+    static int16_t hammingDistance(const string& s1, const string& s2) {
 
         if (s1.size() != s2.size())
             BOOST_THROW_EXCEPTION(SeqUtilsException() << SeqUtilsErrorInfo(string(
@@ -50,7 +50,7 @@ public:
         string s1u = boost::to_upper_copy(s1);
         string s2u = boost::to_upper_copy(s2);
 
-        uint32_t sum = 0;
+        int16_t sum = 0;
         for(size_t i = 0; i < s1u.size(); i++) {
             if (s1u[i] != s2u[i]) {
                 sum++;
